core: be db files refactor
